+++ manifest.json
@@ -3,6 +3,6 @@
   "description": "A chrome extension to intercept and modify HTTP requests",
   "version": "1.0",
   "manifest_version": 3,
-  "devtools_page": "./static/devtools-background.html",
+  "devtools_page": "static/background.html",
   "permissions": ["debugger"]
 }
