--- src/devtools-background.ts
@@ -1,1 +1,0 @@
-chrome.devtools.panels.create('Interceptor', '', 'static/index.html')
