+++ static/background.html
@@ -1,0 +1,2 @@
+<meta charset="utf-8" />
+<script src="../out/background/index.js"></script>
