--- static/devtools-background.html
@@ -1,2 +1,0 @@
-<meta charset="utf-8" />
-<script src="../out/devtools-background.js"></script>
