+++ static/panel.html
@@ -13,6 +13,6 @@
     >
 
     <div id="app"></div>
-    <script src="../out/index.js"></script>
+    <script src="../out/app/index.js"></script>
   </body>
 </html>
